get total working

+++ src/main/java/com/codecool/shop/dao/CartDao.java
@@ -12,8 +12,8 @@ public interface CartDao {
     Integer getTotalValues();
     void addProduct(Product product);
 
-    Double getProductTotal(Product product);
+    float getProductTotal(Product product);
 
-    Double getTotalOfAll();
+    float getTotalOfAll();
 
 }
